pyproject.toml
[project]
name = "dspy-ai-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.1.2",
    "pandas",
    "tqdm",
    "python-dotenv",
    "scikit-learn>=1.8.0",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pyright]
include = ["."]
executionEnvironments = [{root = ".", extraPaths = ["src", "."]}]
